Move tests to their corresponding folders.
